Fix history not saving on windows
The file, the history get's stored is hardcoded which is not a problem in itself.
The problem is, that the path separator is hardcoded, which prevented the script to save the history on Windows which makes the whole script useless.

The script now detects the type of machine it runs on and sets the separator accordingly.