Fix issue with files containing commas
The CSV data file uses commas as separator. The data file contains paths to indexed songs.
It is not common to use commas in files but it can happen. Because of this, some files get loaded
incorrectly and thus not updated across restarts of the extension.

This is now fixed as this version will look for the separator from the end of the line
instead of the beginning where the path is stored.