Fix store not getting loaded properly
The store didn't get properly populated because the path was attempted to be
converted to a number (which doesn't work and returns a nil value)
The subsequent check if path is of any value failed and the sore doesn't get the
new entry.

This is fixed by simply removing the tonumber. Also the date now get's converted
to a number.

The issue was likely introduced in 10c94a93bfdb488a34bb52ebda8f007c4e2de098