Improve randomizing of playlist
The randomizing algorithm was flawed. The randomizing wasn't really be
the probability of the liking. For example if the data file contained
songs with liking of 110 and new songs get added to the data file (now
with liking 100), the newly added songs appear only after most of the
old songs with higher liking.

New behavior is that the songs get added with more diversity. The
example above does not happen (or not as extreme) with the new algorithm.